Adding losing streaks report, re-work urls
