+++ config.dist.json
@@ -10,6 +10,7 @@
 
     "settings": {
         "ga_property_code": null,
-        "site_url": ""
+        "site_url": "",
+        "stats_url": ""
     }
 }
