+++ reports/panelist/__init__.py
@@ -5,7 +5,7 @@
 
 from reports.panelist import (aggregate_scores, appearances_by_year,
                               gender_mix, gender_stats, panelist_vs_panelist,
-                              rankings_summary, stats_summary, win_streaks)
+                              rankings_summary, stats_summary, streaks)
 
 __all__ = [
     "aggregate_scores",
@@ -15,5 +15,5 @@ __all__ = [
     "panelist_vs_panelist",
     "rankings_summary",
     "stats_summary",
-    "win_streaks"
+    "streaks"
     ]
