+++ templates/index.html
@@ -18,7 +18,7 @@
 
 <div id="summary">
     This site provides reports on various aspects of the data recorded in the
-    <a href="//wwdt.me/">Wait Wait... Don't Tell Me! Stats and Show Details</a>
+    <a href="{{ stats_url }}">Wait Wait... Don't Tell Me! Stats and Show Details</a>
     database that don't really fit neatly into the main Stats and Show Details
     site.
 </div>
@@ -143,6 +143,12 @@
             <td>
                 A break down of how well all panelists have done by gender
                 and by year.
+            </td>
+        </tr>
+        <tr>
+            <td><a href="{{ url_for('panelist_losing_streaks') }}">Losing Streaks</a></td>
+            <td>
+                A listing of the longest losing streak for each panelist.
             </td>
         </tr>
         <tr>
