+++ templates/panelist/losing_streaks.html
@@ -1,0 +1,74 @@
+{% extends "base.html" %}
+{% block css %}panelist/losing_streaks.css{% endblock %}
+{% block title %}Panelist Losing Streaks{% endblock %}
+{% block header_title %}Panelist Losing Streaks Report{% endblock %}
+{% block synopsis %}
+<p>
+    Listing of the longest losing streak for each panelist. A loss is defined
+    as not coming in either first place or tied for first.
+</p>
+{% endblock %}
+{% block breadcrumb %}
+<div id="homelink">&lt; <a href="{{ url_for('index') }}">Home</a></div>
+{% endblock %}
+{% block content %}
+<!-- Start Panelist Losing Streaks Section -->
+<table class="pure-table pure-table-bordered">
+    <colgroup>
+        <col class="panelist-name">
+        <col class="panelist-streak-total">
+        <col class="panelist-streak-count">
+        <col class="panelist-streak-dates">
+        <col class="panelist-streak-total">
+        <col class="panelist-streak-count">
+        <col class="panelist-streak-dates">
+    </colgroup>
+    <thead>
+        <tr>
+            <th scope="col" rowspan="2">Panelist</th>
+            <th colspan="3">Losses</th>
+        </tr>
+        <tr>
+            <th scope="col">Total</th>
+            <th scope="col">#</th>
+            <th scope="col">Shows</th>
+        </tr>
+    </thead>
+    <tbody>
+        {% for panelist in losing_streaks %}
+        <tr>
+            <td>
+                <a href="{{ stats_url }}/panelists/{{ panelist.slug }}">
+                    {{ panelist.name }}
+                </a>
+            </td>
+            <td>{{ panelist.total_losses }}</td>
+            <td>{{ panelist.longest_losing_streak }}</td>
+            {% if not panelist.longest_losing_streak_dates %}
+            <td class="no-data">&nbsp;</td>
+            {% else %}
+            <td class="show-dates">
+                <ul>
+                {% for show in panelist.longest_losing_streak_dates %}
+                <li>{{ show.show_date }}</li>
+                {% endfor %}
+                </ul>
+            </td>
+            {% endif %}
+        </tr>
+        {% endfor %}
+    </tbody>
+    <tfoot>
+        <tr>
+            <th scope="col" rowspan="2">Panelist</th>
+            <th scope="col">Total</th>
+            <th scope="col">#</th>
+            <th scope="col">Shows</th>
+        </tr>
+        <tr>
+            <th colspan="3">Loses</th>
+        </tr>
+    </tfoot>
+</table>
+<!-- Start Panelist Losing Streaks Section -->
+{% endblock %}
