+++ templates/panelist/panelist_vs_panelist.html
@@ -32,7 +32,7 @@
                 <span class="pure-menu-heading">Site</span>
                 <ul class="pure-menu-list">
                     <li class="pure-menu-item"><a href="{{ url_for('index') }}" class="pure-menu-link">Home</a></li>
-                    <li class="pure-menu-item"><a href="//wwdt.me/" class="pure-menu-link">Stats Page</a></li>
+                    <li class="pure-menu-item"><a href="{{ stats_url }}" class="pure-menu-link">Stats Page</a></li>
                 </ul>
 
                 <span class="pure-menu-heading">Panelists</span>
