+++ templates/panelist/rankings_summary.html
@@ -37,7 +37,7 @@
     <tbody>
         {% for panelist in panelists_rankings %}
         <tr>
-            <td><a href="https://stats.wwdt.me/panelists/{{ panelist }}">
+            <td><a href="{{ stats_url }}/panelists/{{ panelist }}">
                 {{ panelists[panelist].name }}</a></td>
             {% if panelists_rankings[panelist] %}
             <td>{{ panelists_rankings[panelist]["first"] }}</td>
