+++ templates/panelist/win_streaks.html
@@ -42,7 +42,7 @@
         {% for panelist in win_streaks %}
         <tr>
             <td>
-                <a href="https://stats.wwdt.me/panelists/{{ panelist.slug }}">
+                <a href="{{ stats_url }}/panelists/{{ panelist.slug }}">
                     {{ panelist.name }}
                 </a>
             </td>
