+++ templates/show/lightning_round_end_three_way_tie.html
@@ -29,7 +29,7 @@
     <tbody>
         {% for show in shows %}
         <tr>
-            <td><a href="https://stats.wwdt.me/shows/{{ show.date }}">{{ show.date }}</a></td>
+            <td><a href="{{ stats_url }}/shows/{{ show.date|replace('-', '/') }}">{{ show.date }}</a></td>
             <td class="panelist-names">
                 <ul>
                 {% for panelist in show.panelists %}
